Remove separate grpc test
