+++ .github/workflows/test_key_release_latest.yml
@@ -16,11 +16,3 @@ jobs:
     with:
       example-name: key_release
       manifest-modifiers: "KEY_RELEASE_IMAGE=http://confidentialsidecars.azurecr.io/skr:main"
-
-  key-release-grpc:
-    name: Key Release GRPC
-    uses: ./.github/workflows/test_example.yml
-    secrets: inherit
-    with:
-      example-name: key_release_grpc
-      manifest-modifiers: "KEY_RELEASE_IMAGE=http://confidentialsidecars.azurecr.io/skr:main"
