+++ .vscode/launch.json
@@ -189,7 +189,6 @@
                 "attestation",
                 "remote_image",
                 "key_release",
-                "key_release_grpc",
                 "encrypted_filesystem",
             ],
         },
