+++ examples/key_release/manifest.json
@@ -14,6 +14,17 @@
                     "ram": 2
                 },
                 {
+                    "image": "$KEY_RELEASE_IMAGE",
+                    "env": {
+                        "Port": "50000",
+                        "ServerType": "grpc"
+                    },
+                    "command": [ "/skr.sh" ],
+                    "ports": [ 8081, 50000 ],
+                    "cores": 1,
+                    "ram": 2
+                },
+                {
                     "image": "primary",
                     "env": {
                         "SkrClientMAAEndpoint": "$AZURE_ATTESTATION_ENDPOINT",
@@ -25,7 +36,7 @@
                     "ports": [ 8000 ]
                 }
             ],
-            "ports": [ 8000, 8080 ]
+            "ports": [ 8000, 8080, 8081 ]
         }
     ],
     "securityPolicies": [],
