--- examples/key_release_grpc/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-
-
