--- examples/key_release_grpc/manifest.json
@@ -1,47 +1,0 @@
-{
-    "testName": "key_release_grpc",
-    "images": {
-        "primary": "key_release_grpc/Dockerfile.primary"
-    },
-    "containerGroups": [
-        {
-            "containers": [
-                {
-                    "image": "$KEY_RELEASE_IMAGE",
-                    "env": {
-                        "Port": "50000",
-                        "ServerType": "grpc"
-                    },
-                    "command": [ "/skr.sh" ],
-                    "ports": [ 8080, 50000 ],
-                    "cores": 1,
-                    "ram": 2
-                },
-                {
-                    "image": "primary",
-                    "env": {
-                        "SkrClientMAAEndpoint": "$AZURE_ATTESTATION_ENDPOINT",
-                        "SkrClientAKVEndpoint": "$AZURE_HSM_ENDPOINT",
-                        "SkrClientKID": "$UNIQUE_ID-key"
-                    },
-                    "cores": 1,
-                    "ram": 2,
-                    "ports": [ 8000 ]
-                }
-            ],
-            "ports": [ 8080, 8000, 50000 ]
-        }
-    ],
-    "securityPolicies": [],
-    "registryCredentials": {
-        "$AZURE_REGISTRY_URL": {
-            "username": "$AZURE_REGISTRY_USERNAME",
-            "password": "$AZURE_REGISTRY_PASSWORD"
-        },
-        "$SIDECARS_REGISTRY_URL": {
-            "username": "$SIDECARS_REGISTRY_USERNAME",
-            "password": "$SIDECARS_REGISTRY_PASSWORD"
-        }
-    },
-    "preDeployScript": "deploy_key.py"
-}
